runtime: relocate runtime-typed app_context accessors (Phase 3, Change 3a)
Move hl_app_context_is_lua / _lua / _js out of app_context.c into a new
toolchain-only TU app_context_runtime.c. They compare ctx->factory against
hl_lua_factory / hl_js_factory (concrete runtime symbols); keeping them in the
base app-runner object forced every produced app to drag in BOTH factories, so a
slim single-runtime app could not link. app_context.c now exposes only the
agnostic hl_app_context_factory() getter the accessors build on, so
app_context.o references no concrete runtime symbol (nm confirms: it exports only
hl_app_context_factory; the accessors moved to app_context_runtime.o).

Only the hull toolchain's agent introspection (agent/{eval,template,overview})
calls these. app_context_runtime.o is compiled with both runtime macros; for now
it links beside app_context.o (Change 3b moves it into the toolchain-only object
group, out of the produced-app platform lib).

Byte-identical: make test 60/60 green; hull agent overview/eval work. serve.c's
auto_detect_entry ungating deferred to 3b (entangled with the #ifndef
HL_ENABLE_* runtime validation + a pre-existing JS-first vs Lua-first order).
